Rename test fixture module to expectCallstack.ts
